+++ omechu-app/src/app/mypage/user-info-edit/[step]/AllergyStep.tsx
@@ -7,7 +7,7 @@ import ProgressBar from "@/app/components/common/ProgressBar";
 import ModalWrapper from "@/app/components/common/ModalWrapper";
 import AlertModal from "@/app/components/common/AlertModal";
 import { useOnboardingStore } from "@/lib/stores/onboarding.store";
-import { indexToSlug } from "./page";
+import { indexToSlug } from "@/app/constant/UserInfoEditSteps";
 
 export default function AllergyStep() {
   const router = useRouter();
