+++ omechu-app/src/app/mypage/user-info-edit/[step]/ConditionStep.tsx
@@ -8,7 +8,7 @@ import { useOnboardingStore } from "@/lib/stores/onboarding.store";
 import AlertModal from "@/app/components/common/AlertModal";
 import ModalWrapper from "@/app/components/common/ModalWrapper";
 import ProgressBar from "@/app/components/common/ProgressBar";
-import { indexToSlug } from "./page";
+import { indexToSlug } from "@/app/constant/UserInfoEditSteps";
 
 export default function ConditionStep() {
   const router = useRouter();
